+++ src/calibratedialog.cpp
@@ -16,7 +16,8 @@ CalibrateDialog::CalibrateDialog(QWidget *parent) :
         ui->valueSlider->setValue(static_cast<quint16>(qRound(value*code->getDivider())));
         ui->valueDoubleSpinBox->setValue(value);
     });
-    connect(this, SIGNAL(accepted()), this, SLOT(saveResult()));
+    connect(ui->buttonBox, SIGNAL(accepted()), this, SLOT(saveResult()));
+    connect(ui->buttonBox, SIGNAL(rejected()), this, SLOT(hide()));
 }
 
 CalibrateDialog::~CalibrateDialog()
@@ -28,20 +29,28 @@ void CalibrateDialog::setStruct(const calibration_t &calibrationData) {
     code = calibrationData.cmd;
 
     ui->titleLabel->setText(calibrationData.title);
-    QString desc = QString("Min: %1, Max: %2").arg(QString::number(calibrationData.min)).arg(QString::number(calibrationData.max));
-    ui->valueSlider->setMinimum(calibrationData.min);
-    ui->valueSlider->setMaximum(calibrationData.max);
+    ui->valueSlider->setMinimum(qRound(calibrationData.min*code->getDivider()));
+    ui->valueSlider->setMaximum(qRound(calibrationData.max*code->getDivider()));
     ui->valueDoubleSpinBox->setSingleStep(1.0/code->getDivider());
-    ui->valueDoubleSpinBox->setDecimals(qRound(0.43429*qLn(code->getDivider())));
-    ui->valueDoubleSpinBox->setMinimum(static_cast<double>(calibrationData.min)/code->getDivider());
-    ui->valueDoubleSpinBox->setMaximum(static_cast<double>(calibrationData.max)/code->getDivider());
+    int precision = qRound(0.43429*qLn(code->getDivider()));
+    ui->valueDoubleSpinBox->setDecimals(precision);
+
+    ui->valueDoubleSpinBox->setMinimum(calibrationData.min);
+    ui->valueDoubleSpinBox->setMaximum(calibrationData.max);
+
+    ui->minLabel->setText(wlocale.toString(calibrationData.min, DOUBLE_FORMAT, precision));
+    ui->maxLabel->setText(wlocale.toString(calibrationData.max, DOUBLE_FORMAT, precision));
+
+    ui->valueSlider->setValue(qRound(code->getValue()*code->getDivider()));
+    ui->valueDoubleSpinBox->setValue(code->getValue());
 }
 
 void CalibrateDialog::saveResult() {
     int value = ui->valueSlider->value();
-    QString strToSend = QString("%1%2 %3").arg(COM_WRITE_PREFIX).arg(code->getCode()).arg(value, 4, 16, QChar('0')).toUpper();
+    QString strToSend = QString("%1%2 %3").arg(COM_WRITE_PREFIX).arg(code->getCode()).arg(static_cast<quint16>(value), 4, 16, QChar('0')).toUpper();
     code->resetInterval();
     emit sendData(strToSend);
+    this->hide();
 }