+++ src/signedcommand.h
@@ -10,6 +10,7 @@ class SignedCommand : public Command
 public:
     explicit SignedCommand(QString code, QString unit = "", double divider = 1, quint8 interval = 1, bool isTemperature = false, QObject* parent = nullptr);
     bool isSignedValue() override;
+    int getRawValue() override;
 //    double getValue() override;
 
 public slots: